Replace app icon with original
The original app icon has a wide aspect ratio which doesn't work well
for an app icon, which favor squares. However, I think it like it better
anyway. The original is like the new icon, except the Os form a row
instead of a column and the image is scaled down to fit.